Moves iq learn functions to class level
